Read e00 files

These functions generate temporary binary coverages under the hood and delete them afterward. The overhead for this operation is small, but you can do this conversion yourself using e00_convert().

e00_read_sf(e00, layer = "PAL", ...)

e00_layers(e00, ...)

e00_readOGR(e00, layer = "PAL", ...)

Arguments

e00
Path to a .e00 file.

layer
One of "ARC" (lines), "CNT" (centroids), "LAB" (label points), or "PAL" (polygons).

...
Passed to sf::read_sf(), rgdal::readOGR(), or sf::st_layers().

Examples

e00 <- system.file("e00/cis_20170911.e00", package = "arce00")

e00_layers(e00)

#> Driver: AVCBin 
#> Available layers:
#>   layer_name geometry_type features fields
#> 1        ARC   Line String      230     11
#> 2        CNT         Point       88      1
#> 3        LAB         Point       87     68
#> 4        PAL       Polygon       87     67

e00_read_sf(e00)

#> Simple feature collection with 87 features and 67 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 39656.16 ymin: 1165135 xmax: 2687056 ymax: 3743606
#> projected CRS:  unnamed
#> # A tibble: 87 x 68
#>       AREA PERIMETER HFDVHIYVQGF0F. HFDVHIYVQGF0F.ID A_LEGEND REGION DATE_CARTE
#>      <dbl>     <dbl>          <int>            <int> <chr>    <chr>  <chr>     
#>  1 8.85e11 31235633.              2                2 Bergy w… HB     20170911  
#>  2 1.48e 8    94500.              3                3 Land     HB     20170911  
#>  3 2.24e11 14482767.              4                4 Land     HB     20170911  
#>  4 1.27e 8    53164.              5                5 Land     HB     20170911  
#>  5 1.44e 8    87332.              6                6 Land     HB     20170911  
#>  6 1.10e 8    71738.              7                7 Land     HB     20170911  
#>  7 1.32e 8   110787.              8                8 Land     HB     20170911  
#>  8 8.32e 7    78102.              9                9 Land     HB     20170911  
#>  9 8.66e10  2431778.             10               10 Ice free HB     20170911  
#> 10 4.18e 7    26337.             11               11 Land     HB     20170911  
#> # … with 77 more rows, and 61 more variables: SOURCE <chr>, MOD <chr>,
#> #   EGG_ID <int>, PNT_TYPE <int>, EGG_NAME <chr>, EGG_SCALE <int>,
#> #   EGG_ATTR <chr>, USER_ATTR <chr>, ROTATION <int>, E_CT <chr>, E_CA <chr>,
#> #   E_CB <chr>, E_CC <chr>, E_CD <chr>, E_SO <chr>, E_SA <chr>, E_SB <chr>,
#> #   E_SC <chr>, E_SD <chr>, E_SE <chr>, E_FA <chr>, E_FB <chr>, E_FC <chr>,
#> #   E_FD <chr>, E_FE <chr>, E_CS <chr>, R_CT <chr>, R_CMY <chr>, R_CSY <chr>,
#> #   R_CFY <chr>, R_CGW <chr>, R_CG <chr>, R_CN <chr>, R_PMY <chr>, R_PSY <chr>,
#> #   R_PFY <chr>, R_PGW <chr>, R_PG <chr>, R_PN <chr>, R_CS <chr>, R_SMY <chr>,
#> #   R_SSY <chr>, R_SFY <chr>, R_SGW <chr>, R_SG <chr>, R_SN <chr>, N_CT <chr>,
#> #   N_COI <chr>, N_CMY <chr>, N_CSY <chr>, N_CFY <chr>, N_CFY_TK <chr>,
#> #   N_CFY_M <chr>, N_CFY_TN <chr>, N_CYI <chr>, N_CGW <chr>, N_CG <chr>,
#> #   N_CN <chr>, N_CB <chr>, ArcIds <list>, geometry <POLYGON [m]>

sp <- e00_readOGR(e00, layer = "LAB")

#> OGR data source with driver: AVCBin 
#> Source: "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp22E0DC/file61e730034e4/ivjF0HqZ3vEaf", layer: "LAB"
#> with 87 features
#> It has 68 fields

summary(sp)

#>                 Length                  Class                   Mode 
#>                     87 SpatialPointsDataFrame                     S4